fix(sysctl): Remove panic settings, use values from the kernel.
The default watchdog is 10 seconds, we are using 5 which is perhaps a
bit too aggressive.

The kernel is also configured to delay rebooting after panic for 60
seconds, giving users a chance to see the error first. We don't do
anything to collect any panic info that may be preserved in ram as the
comment here implied so the immediate reboot wasn't helping anyone.